refactor(util): split 1701-line file into module dir
Decomposes src/util.rs into src/util/ (8 files) by responsibility. Pure
code-move: behavior byte-identical, only visibility/import glue added.

- dprint.rs         dprint (11)
- filter.rs         filter_hosts (88)
- hosts_json.rs     parse_hosts_json_{file,url} + private parse_hosts_json (105)
- regex_helpers.rs  regex_is_match (11)
- display.rs        host_display, task_display (16)
- host_info.rs      HostInfo/OSInfo/CPUInfo/MemoryInfo structs + host_info +
                    create_unknown_host_info/parse_host_info_output/
                    create_info_table helpers (292)
- mod.rs            submodule decls + pub use re-exports (22)
- tests.rs          #[cfg(test)] unit tests, moved wholesale (1181)

Cross-submodule refs: filter_hosts -> crate::util::regex_is_match;
parse_hosts_json_* -> crate::util::dprint. Private helpers/structs re-exported
under #[cfg(test)] only (tests call them directly; keeps public API clean).
