+++ src/util/mod.rs
@@ -1,0 +1,22 @@
+mod display;
+mod dprint;
+mod filter;
+mod host_info;
+mod hosts_json;
+mod regex_helpers;
+
+#[cfg(test)]
+mod tests;
+
+#[cfg(test)]
+pub use host_info::{
+    CPUInfo, HostInfo, MemoryInfo, OSInfo, create_info_table, create_unknown_host_info,
+    parse_host_info_output,
+};
+
+pub use display::{host_display, task_display};
+pub use dprint::dprint;
+pub use filter::filter_hosts;
+pub use host_info::host_info;
+pub use hosts_json::{parse_hosts_json_file, parse_hosts_json_url};
+pub use regex_helpers::regex_is_match;
